pyproject.toml
[tool.poetry]
name = "pipecat-project"
version = "0.1.0"
description = ""
authors = ["Amir Hamza <amir@educationequation.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
pipecat-ai = {extras = ["daily", "openai", "silero"], version = "^0.0.49"}
asyncio = "^3.4.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
